Add new method to game and change args

+++ TicTacToe.DAL/DataModels/Step.cs
@@ -3,8 +3,8 @@
     public class Step
     {
         public int Id { get; set; }
-        public byte Row { get; set; }
-        public byte Col { get; set; }
+        public int Row { get; set; }
+        public int Col { get; set; }
         public char Value { get; set; }
     }
 }
